Version 0.13.2: Fix 6 UI/UX issues
Issue 1: Fix slash character input in config textareas

- Allow '/' to be typed in textareas (was blocked by keyboard shortcut)

Issue 2: Fix toast notification positioning

- Simplify positioning to prevent cutoff on right side

- Use inline style right: 1rem instead of conflicting Tailwind classes

Issue 3: Fix stale running scan cleanup

- Automatically mark scans as 'interrupted' if running > 24 hours

- Handles cases where app crashed before updating scan status

Issue 4: Fix .Page template type error

- Change Page field from int to int64 in FilesData, HardlinksData, ScansData

- Cast page values to int64 when creating template data

Issue 5: Add test functionality for scan paths and mappings

- Add test buttons to config page for scan paths and path mappings

- Add HandleTestScanPaths and HandleTestPathMappings handlers

- Validate paths exist and are accessible

- Register /api/config/test-scan-paths and /api/config/test-path-mappings routes

Issue 6: Auto-expand textareas for paths and mappings

- Create textarea-autoresize.js for automatic height adjustment

- Add data-autoresize attribute to config textareas

- Textareas now expand/contract automatically as content changes

+++ internal/server/types.go
@@ -98,7 +98,7 @@ type DashboardData struct {
 type FilesData struct {
 	Files      interface{}
 	Total      int
-	Page       int
+	Page       int64
 	Limit      int
 	TotalPages int
 	Title      string
@@ -124,7 +124,7 @@ type StatsData struct {
 type HardlinksData struct {
 	Groups     interface{}
 	Total      int
-	Page       int
+	Page       int64
 	TotalPages int
 	Title      string
 }
@@ -133,7 +133,7 @@ type HardlinksData struct {
 type ScansData struct {
 	Scans      interface{}
 	Total      int
-	Page       int
+	Page       int64
 	TotalPages int
 	Title      string
 }
